+++ CHANGELOG.markdown
@@ -6,6 +6,8 @@
 * Factored out a common `reversed` definition from all the various forms of it around the library and placed it in `Control.Lens.Iso`.
 * Added `binary, `octal`, `decimal` and `hex` to `Numeric.Lens`.
 * Added `\_Void` to Control.Lens.Void
+* Removed `Gettable`. We now use `Contravariant` and `Functor` together to derive `Getter` and `Fold`. This enables these to be defined
+  without incurring a dependency on the `lens` package.
 
 3.8.5
 -----